pyproject.toml
[tool.poetry]
name = "yt-dl-ocr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "yt_dl_ocr"}]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.58.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
python-dotenv = "^0.20.0"
isort = "^5.10.1"
PySocks = "^1.7.1"
youtube_dl = "^2021.12.17"
opencv-python-headless = "^4.6.0.66"
pytesseract = "^0.3.10"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
